[Country search] Code refactoring
- delite dublicate API service
- add new method to API service / fetchEventByCountryCode
- move refs.cardContainer.innerHTML = '' to function clearCardContainer() / DRY
